# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  routes.TbaEvent:
    properties:
      city:
        type: string
      country:
        type: string
      end_date:
        type: string
      event_code:
        type: string
      event_type:
        type: integer
      key:
        type: string
      name:
        type: string
      start_date:
        type: string
      state_prov:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Blue Banner Engine, providing FRC match
    predictions.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blue Banner Engine (BBE) API
  version: "1.0"
paths:
...
  /events/{year}:
    get:
      description: Retrieves a simplified list of all FRC events for a given year
        from The Blue Alliance.
      parameters:
      - description: FRC Year (e.g., 2024)
        in: path
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.TbaEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get All Events for a Year
      tags:
      - events
  /events/teams/{eventKey}:
    get:
      description: Retrieves a list of all team keys participating in a given FRC
        event from The Blue Alliance.
      parameters:
      - description: FRC Event Key (e.g., 2025mexas)
        in: path
        name: eventKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get All Teams for an Event
      tags:
      - events
...
  /tba/event/{eventKey}/schedule:
    get:
      description: Retrieves the match schedule for an FRC event from The Blue Alliance
      parameters:
      - description: FRC Event Key (e.g., 2025mxmo)
        in: path
        name: eventKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get TBA Event Schedule
      tags:
      - events
  /tba/event/{eventKey}/team/{teamNumber}/matches:
    get:
      description: Retrieves matches for a specific team at an FRC event from The
        Blue Alliance
      parameters:
      - description: FRC Event Key (e.g., 2025mxmo)
        in: path
        name: eventKey
        required: true
        type: string
      - description: Team Number
        in: path
        name: teamNumber
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.ErrorResponse'
      summary: Get Team Event Matches
      tags:
      - events
...
swagger: "2.0"